Reduce CLI binary size by ~5MB
