+++ Dockerfile.twtd
@@ -32,6 +32,7 @@ COPY ./internal/*.go ./internal/
 COPY ./internal/auth/*.go ./internal/auth/
 COPY ./internal/session/*.go ./internal/session/
 COPY ./internal/passwords/*.go ./internal/passwords/
+COPY ./types/*.go ./types/
 COPY ./cmd/twtd/*.go ./cmd/twtd/
 
 # Version/Commit (there there is no .git in Docker build context)
